Migrate agent_memory
