+++ front/components/actions/mcp/details/MCPActionDetails.tsx
@@ -50,11 +50,6 @@ import {
   ENABLE_SKILL_TOOL_NAME,
 } from "@app/lib/actions/constants";
 import {
-  AGENT_MEMORY_COMPACT_TOOL_NAME,
-  AGENT_MEMORY_EDIT_TOOL_NAME,
-  AGENT_MEMORY_ERASE_TOOL_NAME,
-  AGENT_MEMORY_RECORD_TOOL_NAME,
-  AGENT_MEMORY_RETRIEVE_TOOL_NAME,
   DATA_WAREHOUSES_DESCRIBE_TABLES_TOOL_NAME,
   DATA_WAREHOUSES_FIND_TOOL_NAME,
   DATA_WAREHOUSES_LIST_TOOL_NAME,
@@ -92,6 +87,13 @@ import {
   isWebsearchInputType,
 } from "@app/lib/actions/mcp_internal_actions/types";
 import { MCP_SPECIFICATION } from "@app/lib/actions/utils";
+import {
+  AGENT_MEMORY_COMPACT_TOOL_NAME,
+  AGENT_MEMORY_EDIT_TOOL_NAME,
+  AGENT_MEMORY_ERASE_TOOL_NAME,
+  AGENT_MEMORY_RECORD_TOOL_NAME,
+  AGENT_MEMORY_RETRIEVE_TOOL_NAME,
+} from "@app/lib/api/actions/servers/agent_memory/metadata";
 import { isValidJSON } from "@app/lib/utils/json";
 import type { LightWorkspaceType } from "@app/types";
 import { asDisplayName, isSupportedImageContentType } from "@app/types";
