+++ front/lib/actions/mcp_internal_actions/utils.ts
@@ -3,7 +3,6 @@ import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 
 import type { InternalMCPServerNameType } from "@app/lib/actions/mcp_internal_actions/constants";
 import {
-  AGENT_MEMORY_SERVER_NAME,
   getInternalMCPServerInfo,
   matchesInternalMCPServerName,
 } from "@app/lib/actions/mcp_internal_actions/constants";
@@ -72,7 +71,7 @@ export function makeMCPToolExit({
 
 export function isJITMCPServerView(view: MCPServerViewType): boolean {
   return (
-    !matchesInternalMCPServerName(view.server.sId, AGENT_MEMORY_SERVER_NAME) &&
+    !matchesInternalMCPServerName(view.server.sId, "agent_memory") &&
     // Only tools that do not require any configuration can be enabled directly in a conversation.
     getMCPServerRequirements(view).noRequirement
   );
